Add files via upload
AccelerEQ is a Max project that utilizes the Microbit's accelerometer to make complex creative EQing in production and a live setting more dynamic and accessible. It was created as an assignment submission for the Interfacing for Audio & Music module as part of the Audio and Music Technology MSc program at the University of York.